rename: 파일 구조 변경 & import 경로 변경
